+++ src/main/java/com/unzer/payment/service/PaymentService.java
@@ -107,6 +107,8 @@ public class PaymentService {
         switch (paymentTypeEnum) {
             case CARD:
                 return new Card("", "");
+            case CLICK_TO_PAY:
+                return new ClickToPay();
             case EPS:
                 return new Eps();
             case GIROPAY:
@@ -190,6 +192,7 @@ public class PaymentService {
             case PF_EFINANCE:
             case UNZER_PAYLATER_INVOICE:
             case KLARNA:
+            case CLICK_TO_PAY:
                 return new ApiIdObject();
             case PAYPAL:
                 return new ApiPaypal();