+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 * Fixed `Unzer:updateCharge` and `Unzer:updateAuthorization` uses wrong endpoints
+* SEPA Direct Debit is restored from deprecation. See: `com.unzer.payment.paymenttypes.SepaDirectDebit`
 
 ## [5.0.0](https://github.com/unzerdev/java-sdk/compare/4.4.0..5.0.0)