+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
 ## [5.1.0](https://github.com/unzerdev/java-sdk/compare/5.0.0..5.1.0)
 
 Google Pay payment method is added to Java SDK.