+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Click To Pay payment method is added to Java SDK.
 ### Added
 
 * Added `ClickToPay` payment method.
-
+* Added `Twint` payment method.
 
 ## [5.1.0](https://github.com/unzerdev/java-sdk/compare/5.0.0..5.1.0)