+++ CHANGELOG.md
@@ -8,10 +8,11 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## UNRELEASED
 
-Twint payment method is added to Java SDK.
+Click To Pay payment method is added to Java SDK.
 
 ### Added
 
+* Added `ClickToPay` payment method.
 * Added `Twint` payment method.
 
 ## [5.1.0](https://github.com/unzerdev/java-sdk/compare/5.0.0..5.1.0)